test(cli): split preview basic tests
